+++ pyrosm/delta_compression.pyx
@@ -6,7 +6,7 @@ cdef delta_decode_latitude(data, node_granularity, lat_offset):
     lats_deltas = np.zeros(len(data.lat) + 1, dtype=np.int64)
     lats_deltas[1:] = list(data.lat)
     lats = (np.cumsum(lats_deltas)[1:] * node_granularity + lat_offset) / div
-    return lats
+    return lats.astype(np.float32)
 
 
 cdef delta_decode_longitude(data, node_granularity, lon_offset):
@@ -14,7 +14,7 @@ cdef delta_decode_longitude(data, node_granularity, lon_offset):
     lons_deltas = np.zeros(len(data.lon) + 1, dtype=np.int64)
     lons_deltas[1:] = list(data.lon)
     lons = (np.cumsum(lons_deltas)[1:] * node_granularity + lon_offset) / div
-    return lons
+    return lons.astype(np.float32)
 
 
 cdef delta_decode_id(data):
@@ -25,18 +25,18 @@ cdef delta_decode_id(data):
 
 
 cdef delta_decode_timestamp(data, timestamp_granularity):
-    timestamp_deltas = np.zeros(len(data.denseinfo.timestamp) + 1, dtype=np.int64)
+    timestamp_deltas = np.zeros(len(data.denseinfo.timestamp) + 1, dtype=np.uint32)
     timestamp_deltas[1:] = list(data.denseinfo.timestamp)
     timestamps = (np.cumsum(timestamp_deltas)[1:]
                   * timestamp_granularity / 1000).astype(int)
-    return timestamps
+    return timestamps.astype(np.uint32)
 
 
 cdef delta_decode_changeset(data):
-    changeset_deltas = np.zeros(len(data.denseinfo.changeset) + 1, dtype=np.int64)
+    changeset_deltas = np.zeros(len(data.denseinfo.changeset) + 1, dtype=np.int32)
     changeset_deltas[1:] = list(data.denseinfo.changeset)
     changesets = np.cumsum(changeset_deltas)[1:]
-    return changesets
+    return changesets.astype(np.int32)
 
 
 cdef delta_encode_latitude(lat_array, node_granularity, lat_offset):
